scripts/requirements.txt
python-igraph
numpy
prettytable
pyyaml
pyllist
cerberus
tqdm
matplotlib
pycairo
seaborn
poverty
haversine
python-tsp
